learning-marketplace/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-community
langgraph
jinja2
chromadb
sentence-transformers
PyPDF2
pdfplumber
pytesseract
python-pptx
Pillow
spacy
networkx
transformers
torch
youtube-transcript-api
openai
